Added: Unit Converter

+++ Components/UnitConverter.php
@@ -200,10 +200,6 @@ class UnitConverter
     {
         //====================================================================//
         // Convert Source to Base Unit then Apply Target Factor
-//        if (extension_loaded("bcmath")) {
-//            return (float) bcdiv( bcmul($value, $toFactor, 9), $fromFactor, 9);
-//        }
-//
-        return (float) ($value * ($toFactor / $fromFactor));
+        return $value * ($toFactor / $fromFactor);
     }
 }
